+++ pages/_app.tsx
@@ -2,21 +2,6 @@ import '../styles/globals.css';
 import type { AppProps } from 'next/app';
 import firebase from 'firebase/app';
 
-const firebaseConfig = {
-  apiKey: process.env.API_KEY,
-  authDomain: process.env.AUTH_DOMAIN,
-  projectId: process.env.PROJECT_ID,
-  storageBucket: process.env.STORAGE_BUCKET,
-  messagingSenderId: process.env.MESSAGING_SENDER_ID,
-  appId: process.env.APP_ID,
-  measurementId: process.env.MEASUREMENT_ID,
-};
-
-if (firebase.apps.length === 0) {
-  // Initialize Firebase
-  firebase.initializeApp(firebaseConfig);
-}
-
 function MyApp({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />;
 }
